整理依赖管理，mbg增加lombok注解插件

+++ mall-admin/src/main/java/com/bob/mall/admin/mapper/UmsPermissionMapper.java
@@ -2,15 +2,19 @@ package com.bob.mall.admin.mapper;
 
 import com.bob.mall.admin.model.UmsPermission;
 import com.bob.mall.admin.model.UmsPermissionExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
+/**
+* Created by Mybatis Generator on 2019/09/08
+*/
 public interface UmsPermissionMapper {
     long countByExample(UmsPermissionExample example);
 
     int deleteByExample(UmsPermissionExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(UmsPermission record);
 
@@ -18,7 +22,7 @@ public interface UmsPermissionMapper {
 
     List<UmsPermission> selectByExample(UmsPermissionExample example);
 
-    UmsPermission selectByPrimaryKey(Integer id);
+    UmsPermission selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") UmsPermission record, @Param("example") UmsPermissionExample example);
