+++ mall-admin/src/main/java/com/bob/mall/admin/mapper/UmsUserRoleRelationMapper.java
@@ -2,15 +2,19 @@ package com.bob.mall.admin.mapper;
 
 import com.bob.mall.admin.model.UmsUserRoleRelation;
 import com.bob.mall.admin.model.UmsUserRoleRelationExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
+/**
+* Created by Mybatis Generator on 2019/09/08
+*/
 public interface UmsUserRoleRelationMapper {
     long countByExample(UmsUserRoleRelationExample example);
 
     int deleteByExample(UmsUserRoleRelationExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(UmsUserRoleRelation record);
 
@@ -18,7 +22,7 @@ public interface UmsUserRoleRelationMapper {
 
     List<UmsUserRoleRelation> selectByExample(UmsUserRoleRelationExample example);
 
-    UmsUserRoleRelation selectByPrimaryKey(Integer id);
+    UmsUserRoleRelation selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") UmsUserRoleRelation record, @Param("example") UmsUserRoleRelationExample example);
