+++ mall-mbg/src/main/java/com/bob/mall/admin/mapper/UmsPermissionMapper.java
@@ -10,7 +10,7 @@ public interface UmsPermissionMapper {
 
     int deleteByExample(UmsPermissionExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(UmsPermission record);
 
@@ -18,7 +18,7 @@ public interface UmsPermissionMapper {
 
     List<UmsPermission> selectByExample(UmsPermissionExample example);
 
-    UmsPermission selectByPrimaryKey(Integer id);
+    UmsPermission selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") UmsPermission record, @Param("example") UmsPermissionExample example);
