+++ mall-mbg/src/main/java/com/bob/mall/admin/mapper/UmsRolePermissionRelationMapper.java
@@ -10,7 +10,7 @@ public interface UmsRolePermissionRelationMapper {
 
     int deleteByExample(UmsRolePermissionRelationExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(UmsRolePermissionRelation record);
 
@@ -18,7 +18,7 @@ public interface UmsRolePermissionRelationMapper {
 
     List<UmsRolePermissionRelation> selectByExample(UmsRolePermissionRelationExample example);
 
-    UmsRolePermissionRelation selectByPrimaryKey(Integer id);
+    UmsRolePermissionRelation selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") UmsRolePermissionRelation record, @Param("example") UmsRolePermissionRelationExample example);
